requirements.txt
absl-py==2.1.0
accelerate==1.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
anyio==4.6.2.post1
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.1.0
babel==2.16.0
beautifulsoup4==4.12.2
bleach==6.2.0
boltons==23.0.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
compel==2.0.3
conda==23.9.0
conda-build==24.3.0
conda-content-trust==0.2.0
conda_index==0.4.0
conda-libmamba-solver==23.7.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.3.0
controlnet-aux==0.0.9
cryptography==42.0.5
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.8
distro==1.8.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.0
executing==0.8.3
expecttest==0.2.1
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.1.2
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hypothesis==6.99.13
idna==3.4
imageio==2.36.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
libarchive-c==2.9
libmambapy==1.5.3
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
menuinst==2.0.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==6.4.4
notebook_shim==0.2.4
numpy==1.26.4
opencv-python-headless==4.10.0.84
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
pip==23.3.1
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.0.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.29.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.0.0
pycosat==0.6.6
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.15.1
pyOpenSSL==24.0.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.30.2
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==68.2.2
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.2.0
stringzilla==3.11.3
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.12
tinycss2==1.4.0
tokenizers==0.20.1
tomli==2.0.1
toolz==0.12.0
torch==2.2.2
torchaudio==2.2.2
torchelastic==0.2.2
torchmetrics==1.6.1
torchvision==0.17.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.7.1
transformers==4.46.1
triton==2.2.0
truststore==0.8.0
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.1.0
wandb==0.19.1
wcwidth==0.2.5
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.41.2
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.0
zipp==3.20.2
zstandard==0.19.0
